memory: Add another missing 32->64 bits conversion when building the invsubmask (nw)

+++ src/emu/memory.c
@@ -4429,7 +4429,7 @@ void handler_entry::clear_conflicting_subunits(UINT64 handlermask)
 	// compute the inverse mask
 	m_invsubmask = 0;
 	for (int i = 0; i < m_subunits; i++)
-		m_invsubmask |= m_subunit_infos[i].m_mask << m_subunit_infos[i].m_shift;
+		m_invsubmask |= UINT64(m_subunit_infos[i].m_mask) << m_subunit_infos[i].m_shift;
 	m_invsubmask = ~m_invsubmask;
 }
